requirements.txt
streamlit>=1.38.0
streamlit-option-menu>=0.3.12
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.9.0
pillow>=10.0.0
requests>=2.31.0
